refactor: inherit interface (IAccount) (#82)
## Summary

Update Account contract to inherit `IAccount`, and update test files

## Details
See comments discussion here: 
* https://github.com/lyra-finance/v2-core/pull/41#discussion_r1057996118

We initially decided to not inherit interfaces because we don't want to
constantly changing 2 files simultaneously while developing. But as we
code more, it seems like the sentiment has changed a bit and we start to
see some good things of this pattern:

* We can declare variables clearly as `IAccount` in contracts easily,
without the need to convert from Account to address and IAccount
* Remove events and errors from the implementation contract, and move to
the interface file (so we don't have to argue about whether they should
be following the standard to come before all functions or after)
* We initially have to duplicate events and errors in both interfaces
and implementation contracts

### New proposed pattern (build cycle):
* Each contract has an empty interface from the very beginning, whenever
others want to call the contract, they can important the interface and
add the "function" they need to call to the interface file. After this
step, the implementation will be "forced" to inform the interface the
change of any function signature, and therefore need to update all
contracts that already integrate with it. (update unit tests to a
different mock call ... etc)

* Interface files are only for easier cooperation for now, we don't need
to spend too much time make sure the documentation is synced during the
building phase. We also don't have to add functions to the interfaces
file if no one else is using it.

* We still use a `Struct` file to declare all structures. So that if
people wants, they can easy say
```solidity
contract AccountTest is AccountStructs {
...
   // use the struct directly
   AssetDelta memory = ...
}
```
instead of `IAccount.AssetDelta` all the time

## Checklist

Ensure you completed **all of the steps** below before submitting your
pull request:

- [x] Add
[natspec](https://docs.soliditylang.org/en/latest/natspec-format.html)
for all functions / parameters
- [x] Ran `forge snapshot`
- [x] Ran `forge fmt`
- [x] Ran `forge test`
- [x] [Triage Slither issues](../README.md#triage-issues), and post
uncertain ones in the PR
- [x] 100% test coverage on code changes

Co-authored-by: Joshua Kim <joshua0702@gmail.com>
Co-authored-by: Joshua Kim <46257136+joshpwrk@users.noreply.github.com>