Feat (Cash Asset): Accrue interest (#95)
Updates totalBorrow/supply with interest accrued over an elapsed period
of time. This change updates both borrow/supply indexes and user
balances are adjusted in the hook on a balance change.

- Inside the hook interest is updated first and applied to the
preBalance before the adjustment is made
- if totalBorrow is 0, we don't enter the interest logic and the
lastTimestamp of interest accrual is updated

Breakdown the progress of the PR for everyone to see what else you
intend to include in this PR.

- [x] state variable for users previous borrow index
- [x] Implementation for `accrueInterest`
- [x] Apply interest to users balance inside hook