Feat: min static balance + other todos (#155)
## Summary

Adding a static margin cash offset to PCRM.InitialMargin to ensure
liquidations are always profitable and to reduce issues with sybil
attacks. As MLRM can never be liquidated, static margin offset is not
required.

The static margin only applies for initialMargin && when there are
options held (both long and short).

Also adding several sprint todo (see below)

## Details

The minimum balance affects the SecurityModule, as someone can block
socialized losses by sending a long option to the SM, thereby forcing it
to hold a minimum $50 balance.

There are several options to deal with this:
- Make Option.sol require allowances if sent to SM
- Only apply static balance if portfolio has shorts
- Always send up to cashBalance - $50 whenever socializing losses

Trade-offs:
- Option A: quite clunky, as it creates a new link between SM <>
Option.sol. Furthermore, every time we redeploy new SMs or Options we
need to not forget to enforce this rule.
- Option B: requires much more gas and big refactor to return some
boolean that says "hasShorts". Still leaves the door open to potential
exploits if somehow we accidentally allow short options to be moved to
the SM.
- Option C: needs a bit of refactoring to force SM to store the PCRM
address and the manager to be IPCRM compliant. However, since the two
are so intertwined, this does not seem to add any extra complexity.

Going with Option C.

## Todo

- [x] Min Static Margin implementation
- [x] Rename PCRMGrouping to StrikeGrouping
- [x] Move OI Fee charging params to BaseManager
- [x] Add manager whitelisting 
- [x] Unit tests

## Checklist

Ensure you completed **all of the steps** below before submitting your
pull request:

- [x] Add
[natspec](https://docs.soliditylang.org/en/latest/natspec-format.html)
for all functions / parameters
- [x] Ran `forge snapshot`
- [x] Ran `forge fmt`
- [x] Ran `forge test`
- [x] [Triage Slither issues](../README.md#triage-issues), and post
uncertain ones in the PR
- [x] 100% test coverage on code changes

### Slither Issues (Optional)

If you're unsure about a new issue reported by Slither, copy them here
so others can verify as well.